requirements.txt
scikit-learn
pandas
numpy
pyarrow
lightgbm
#catboost
numerapi
python-dotenv